fix project and chip width
