+++ src/pages/pegase/home/components/HomePageContent.tsx
@@ -34,7 +34,7 @@ const HomePageContent = () => {
   return (
     <div className="flex w-full flex-1 flex-col justify-start gap-4">
       <StdHeading title={t('home.@study_table_title')} />
-      <div className="flex w-1/3 items-center gap-4">
+      <div className="flex w-2/5 items-center gap-4">
         <div className="flex w-3/4">
           <Searchbar onChange={searchStudy} onSearch={searchStudy} label={t('home.@search_placeholder_study')} />
         </div>
